Extracted k8s delete related code to k8s-delete-utils.js
