+++ helpers.js
@@ -8,115 +8,6 @@ const {
 const HTTP_CODE_UNAUTHORIZED = 401;
 const UNAUTHORIZED_ERROR_MESSAGE = "Please ensure the Service Account Token is vaulted in the correct format and that your service account has sufficient privileges to perform this operation. Consult the plugin documentation for more details.";
 
-const deleteFunctionNamesToApiMap = new Map([
-  ["deleteNamespacedConfigMap", k8s.CoreV1Api],
-  ["deleteNamespacedEndpoints", k8s.CoreV1Api],
-  ["deleteNamespacedEvent", k8s.CoreV1Api],
-  ["deleteNamespacedLimitRange", k8s.CoreV1Api],
-  ["deleteNamespace", k8s.CoreV1Api],
-  ["deleteNode", k8s.CoreV1Api],
-  ["deleteNamespacedPersistentVolumeClaim", k8s.CoreV1Api],
-  ["deleteNamespacedPod", k8s.CoreV1Api],
-  ["deleteNamespacedPodTemplate", k8s.CoreV1Api],
-  ["deleteNamespacedReplicationController", k8s.CoreV1Api],
-  ["deleteNamespacedResourceQuota", k8s.CoreV1Api],
-  ["deleteNamespacedSecret", k8s.CoreV1Api],
-  ["deleteNamespacedServiceAccount", k8s.CoreV1Api],
-  ["deleteNamespacedService", k8s.CoreV1Api],
-  ["deleteNamespacedDaemonSet", k8s.AppsV1Api],
-  ["deleteNamespacedDeployment", k8s.AppsV1Api],
-  ["deleteNamespacedReplicaSet", k8s.AppsV1Api],
-  ["deleteNamespacedStatefulSet", k8s.AppsV1Api],
-  ["deleteNamespacedControllerRevision", k8s.AppsV1Api],
-  ["deleteNamespacedIngress", k8s.NetworkingV1Api],
-  ["deleteNamespacedNetworkPolicy", k8s.NetworkingV1Api],
-  ["deletePodSecurityPolicy", k8s.PolicyV1beta1Api],
-  ["deleteClusterRoleBinding", k8s.RbacAuthorizationV1Api],
-  ["deleteClusterRole", k8s.RbacAuthorizationV1Api],
-  ["deleteClusterRoleBinding", k8s.RbacAuthorizationV1Api],
-  ["deleteNamespacedJob", k8s.BatchV1Api],
-  ["deleteNamespacedCronJob", k8s.BatchV1Api],
-  ["deleteStorageClass", k8s.StorageV1Api],
-  ["deleteVolumeAttachment", k8s.StorageV1Api],
-]);
-
-function mapResourceTypeToDeleteFunctionName(resourceType) {
-  switch (resourceType) {
-    case "configmaps": case "configmap": case "cm":
-      return "deleteNamespacedConfigMap";
-    case "endpoints": case "endpoint": case "ep":
-      return "deleteNamespacedEndpoints";
-    case "events": case "event": case "ev":
-      return "deleteNamespacedEvent";
-    case "limitranges": case "limitrange": case "limits":
-      return "deleteNamespacedLimitRange";
-    case "namespaces": case "namespace": case "ns":
-      return "deleteNamespace";
-    case "nodes": case "node": case "no":
-      return "deleteNode";
-    case "persistentvolumeclaims": case "persistentvolumeclaim": case "pvc":
-      return "deleteNamespacedPersistentVolumeClaim";
-    case "persistentvolumes": case "persistentvolume": case "pv":
-      return "deleteNamespacedPersistentVolume";
-    case "pods": case "pod": case "po":
-      return "deleteNamespacedPod";
-    case "podtemplates": case "podtemplate":
-      return "deleteNamespacedPodTemplate";
-    case "replicationcontrollers": case "replicationcontroller": case "rc":
-      return "deleteNamespacedReplicationController";
-    case "resourcequotas": case "resourcequota": case "quota":
-      return "deleteNamespacedResourceQuota";
-    case "secrets": case "secret":
-      return "deleteNamespacedSecret";
-    case "serviceaccounts": case "serviceaccount": case "sa":
-      return "deleteNamespacedServiceAccount";
-    case "services": case "service": case "svc":
-      return "deleteNamespacedService";
-    case "daemonsets": case "daemonset": case "ds":
-      return "deleteNamespacedDaemonSet";
-    case "deployments": case "deployment": case "deploy":
-      return "deleteNamespacedDeployment";
-    case "statefulsets": case "statefulset": case "sts":
-      return "deleteNamespacedStatefulSet";
-    case "controllerrevisions": case "controllerrevision":
-      return "deleteNamespacedControllerRevision";
-    case "replicasets": case "replicaset": case "rs":
-      return "deleteNamespacedReplicaSet";
-    case "ingresses": case "ingress": case "ing":
-      return "deleteNamespacedIngress";
-    case "networkpolicies": case "networkpolicy": case "netpol":
-      return "deleteNamespacedNetworkPolicy";
-    case "podsecuritypolicies": case "podsecuritypolicy": case "psp":
-      return "deletePodSecurityPolicy";
-    case "clusterrolebindings": case "clusterrolebinding":
-      return "deleteClusterRoleBinding";
-    case "clusterroles": case "clusterrole":
-      return "deleteClusterRole";
-    case "rolebindings": case "rolebinding":
-      return "deleteClusterRoleBinding";
-    case "jobs": case "job":
-      return "deleteNamespacedJob";
-    case "cronjobs": case "cronjob": case "cj":
-      return "deleteNamespacedCronJob";
-    case "storageclasses": case "storageclass": case "sc":
-      return "deleteStorageClass";
-    case "volumeattachments": case "volumeattachment":
-      return "deleteVolumeAttachment";
-    default:
-      throw new Error("Unrecognized resource type");
-  }
-}
-
-function getDeleteApi(resourceType, functionName) {
-  const api = deleteFunctionNamesToApiMap.get(functionName);
-
-  if (!api) {
-    throw new Error(`Couldn't find API client for resource type '${resourceType}'`);
-  }
-
-  return api;
-}
-
 function createK8sClient(api, {
   kubeCertificate,
   kubeApiServer,
@@ -209,8 +100,6 @@ function decodeBase64(content) {
 }
 
 module.exports = {
-  mapResourceTypeToDeleteFunctionName,
-  getDeleteApi,
   createK8sClient,
   applyBySpec,
   parseError,
